requirements.txt
Flask==3.1.0
web3==7.8.0
flask_sqlalchemy==3.1.1
flask_jwt_extended==4.7.1
Flask-Cors==3.0.10